user registration

+++ base/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from .models import Room, Topic
+from django.http import HttpResponse
 from django.db.models import Q
 from .forms import RoomForm
 from django.contrib.auth.models import User
@@ -10,6 +11,10 @@ from django.contrib.auth.decorators import login_required
 # Create your views here.
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -40,7 +45,12 @@ def home(request):
     rooms = Room.objects.filter(Q(topic__name__icontains=q) | Q(name__icontains=q) | Q(description__icontains=q))
     topic = Topic.objects.all()
     room_count = rooms.count()
-    context = {"rooms": rooms, "topics": topic, "room_count": room_count}
+
+    context = {
+        "rooms": rooms, 
+        "topics": topic,
+        "room_count": room_count,
+    }
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
@@ -62,19 +72,31 @@ def createRoom(request):
     }
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
             form.save()
             return redirect("home")
-    context = {'form': form}
+    context = {
+        'form': form,
+    }
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
